Revert "Feat/drag select"
