+++ builder/blocks/prevInception.js
@@ -1,4 +1,7 @@
+import { RedirectWrapper } from './text'
+
 export const PrevInception = (props) => {
+  const redirectUrl = props?.redirect
   const gradientColor = props.gradientColor
 
   let background
@@ -8,36 +11,39 @@ export const PrevInception = (props) => {
     background = props.backgroundColor
   }
   const inceptionModifiers = {
+    // boxShadow: props.boxShadow,
     ...(background ? { background } : {}),
     borderRadius: props.borderRadius,
     border: props.border,
   }
   return (
-    <div
-      style={{
-        width: '100%',
-        height: '100%',
-        margin: 'auto',
-        backgroundImage: `url(${props.imageUrl})`,
-        backgroundPosition: '50% 50%',
-        backgroundRepeat: 'no-repeat',
-        backgroundSize: 'cover',
-        ...inceptionModifiers,
-      }}
-    >
+    <RedirectWrapper redirectUrl={redirectUrl}>
       <div
         style={{
           width: '100%',
           height: '100%',
-          backgroundColor: '#000000',
-          position: 'absolute',
-          top: 0,
-          left: 0,
-          opacity: props.opacity || 0,
-          overflow: 'hidden',
-          borderRadius: props.borderRadius,
+          margin: 'auto',
+          backgroundImage: `url(${props.imageUrl})`,
+          backgroundPosition: '50% 50%',
+          backgroundRepeat: 'no-repeat',
+          backgroundSize: 'cover',
+          ...inceptionModifiers,
         }}
-      />
-    </div>
+      >
+        <div
+          style={{
+            width: '100%',
+            height: '100%',
+            backgroundColor: '#000000',
+            position: 'absolute',
+            top: 0,
+            left: 0,
+            opacity: props.opacity || 0,
+            overflow: 'hidden',
+            borderRadius: props.borderRadius,
+          }}
+        />
+      </div>
+    </RedirectWrapper>
   )
 }
