+++ builder/web-builder/web-builder.js
@@ -1,11 +1,12 @@
-import React, { useCallback, useEffect, useRef, forwardRef } from 'react'
+import React, { useCallback, useEffect, useRef, useState } from 'react'
 import PropTypes from 'prop-types'
-import { Box } from '@chakra-ui/layout'
+import { Box, forwardRef } from '@chakra-ui/react'
+import Draggable from 'react-draggable'
 import { v4 as uuid } from 'uuid'
 
 import { getParentBlock, highlightFutureParentBlock } from './helpers'
 import { GRID_COLUMNS } from './constants'
-import { batch, useDispatch, useSelector } from 'react-redux'
+import { useDispatch, useSelector } from 'react-redux'
 import {
   getGridRowHeight,
   setGridRowHeight,
@@ -13,14 +14,32 @@ import {
   addNewBlock,
   getLayoutsKeys,
   getHierarchy,
+  getBlockLayoutById,
+  handleDragStop,
+  handleResizeStop,
+  handleDrag,
   getNewBlockType,
+  getBlockData,
+  getSelectedBlockId,
   getLayout,
   getIsMobileBuilder,
-  setGroupSelectedBlocksIds,
 } from '../../features/builderSlice'
+import { BuilderBlock, ResizingCounter } from '../blocks'
+import { RayTracing } from './Ray-Tracing'
+import ResizeWrapper from './resizable-wrapper'
 
-import MemoDrag from './components/draggable-item'
-import MultipleSelection from './multiple-selection'
+const blocksZIndex = {
+  inception: 0,
+  image: 1,
+  form: 2,
+  button: 3,
+  text: 4,
+}
+
+function getZIndexValue(blockType, isSelected) {
+  if (isSelected && blockType === 'text') return '4'
+  return blocksZIndex[blockType]?.toString() ?? '2'
+}
 
 const blockSizes = {
   form: { w: 60, h: 10 },
@@ -33,6 +52,144 @@ const blockSizes = {
 function getDroppedBlockDim(blockType) {
   return blockSizes[blockType] || { w: 10, h: 10 }
 }
+
+const DraggableItem = ({
+  blockId,
+  handleHiglightSection,
+  removeHighlightedElem,
+  gridColumnWidth,
+  builderRef,
+}) => {
+  const [resizeValues, setResizeValues] = useState(null)
+  const dispatch = useDispatch()
+  const gridRowHeight = useSelector(getGridRowHeight)
+  const blockLayout = useSelector(getBlockLayoutById(blockId))
+  const selectedBlock = useSelector(getSelectedBlockId)
+  if (!blockLayout) return null
+  const blockData = useSelector(getBlockData(blockId))
+  const isMobile = useSelector(getIsMobileBuilder)
+  const { x, y, w, h } = blockLayout
+  const width = gridColumnWidth * w
+  const height = gridRowHeight * h
+  const xPos = gridColumnWidth * x
+  const yPos = gridRowHeight * y
+  const blockType = blockData?.type
+  const isTextBlock = blockType === 'text'
+  const isSelected = selectedBlock === blockId
+  const [blockPostRef, setBlockPostRef] = useState(null)
+  function onDragStop(_, blockPos) {
+    setBlockPostRef(null)
+    dispatch(handleDragStop(blockPos, blockId))
+    removeHighlightedElem()
+  }
+
+  function onResizeStop(_, __, ___, delta) {
+    dispatch(handleResizeStop(delta, blockId, blockType))
+    setResizeValues(null)
+  }
+
+  function handleResize(_, __, elRef) {
+    const { width, height } = elRef.getBoundingClientRect()
+    setResizeValues({ width: Math.round(width), height: Math.round(height) })
+  }
+
+  function onDrag(_, blockPos) {
+    setBlockPostRef({
+      ...blockPos,
+      w: w * gridColumnWidth,
+      h: h * gridRowHeight,
+      blockId,
+      isDragging: true,
+    })
+    const newBlockLayout = {
+      x: blockPos.x / gridColumnWidth,
+      y: blockPos.y / gridRowHeight,
+      w: w,
+      h: h,
+      i: blockId,
+    }
+    handleHiglightSection(newBlockLayout)
+
+    if (blockId.includes('inception')) {
+      dispatch(
+        handleDrag(
+          blockPos,
+          newBlockLayout,
+          blockId,
+          gridColumnWidth,
+          gridRowHeight
+        )
+      )
+    }
+  }
+
+  const el = document.getElementById(blockId)
+  if (el?.offsetParent) {
+    el.offsetParent.offsetParent.style.zIndex = getZIndexValue(
+      blockType,
+      isSelected
+    )
+  }
+
+  const zIndexValue = getZIndexValue(blockType, isSelected)
+  const isDragging = blockPostRef?.isDragging
+  const right = isMobile ? MobileWindowWidth : window.innerWidth
+  return (
+    <>
+      <Draggable
+        key={blockId}
+        position={{ x: xPos, y: yPos }}
+        onStop={onDragStop}
+        onDrag={onDrag}
+        handle=".draggHandle"
+        bounds={{
+          left: 0,
+          top: 0,
+          right: right - width,
+          bottom: '100% ',
+        }}
+      >
+        <Box pos="absolute" zIndex={zIndexValue}>
+          <ResizeWrapper
+            width={width}
+            height={height}
+            blockId={blockId}
+            onResizeStop={onResizeStop}
+            isTextBlock={isTextBlock}
+            handleResize={handleResize}
+            isSelected={isSelected}
+          >
+            <RayTracing
+              width={width}
+              gridColumnWidth={gridColumnWidth}
+              blockPostRef={blockPostRef}
+              blockId={blockId}
+              isDragging={isDragging}
+              builderRef={builderRef}
+            />
+
+            <MemoBlockItem
+              blockId={blockId}
+              isDragging={isDragging}
+              zIndexValue={zIndexValue}
+            />
+            <ResizingCounter {...resizeValues} pos={{ x: xPos, y: yPos }} />
+          </ResizeWrapper>
+        </Box>
+      </Draggable>
+    </>
+  )
+}
+
+const BlockItem = ({ blockId, isDragging, zIndexValue }) => {
+  return (
+    <Box w={'100%'} h={'100%'} pos="absolute" zIndex={zIndexValue}>
+      <BuilderBlock blockId={blockId} isDragging={isDragging} />
+    </Box>
+  )
+}
+const MemoBlockItem = React.memo(BlockItem)
+const MemoDrag = React.memo(DraggableItem)
 
 function getFontSize(isMobile) {
   if (isMobile) return 10
@@ -51,13 +208,11 @@ const GridLayoutWrapper = forwardRef(
         w="100%"
         height="100%"
         flexDir="row"
-        onClick={() => {
-          dispatch(setBlockEditable(null))
-        }}
+        onClick={() => dispatch(setBlockEditable(null))}
         id="main-builder"
         overflow="hidden"
         pos="relative"
-        className="droppable-element elements selecto-area"
+        className="droppable-element"
         onDragOver={higlightOnDrop}
         bg="gray.50"
         onDrop={(e) => {
@@ -78,7 +233,6 @@ const GridLayoutWrapper = forwardRef(
 GridLayoutWrapper.propTypes = {
   children: PropTypes.any,
 }
-GridLayoutWrapper.displayName = 'GridLayoutWrapper'
 
 export const MobileWindowWidth = 370
 
@@ -92,7 +246,6 @@ const WebBuilder = () => {
   const lastHoveredEl = useRef()
   const gridRowHeight = useSelector(getGridRowHeight)
   const isMobile = useSelector(getIsMobileBuilder)
-
   const windowWidth = isMobile ? MobileWindowWidth : window?.innerWidth
   const columns = isMobile ? GRID_COLUMNS / 2 : GRID_COLUMNS
   const gridColumnWidth = windowWidth / columns
@@ -167,31 +320,25 @@ const WebBuilder = () => {
 
   function handleKeyPress(e) {
     if (e.key === 'Escape') {
-      batch(() => {
-        dispatch(setBlockEditable(null))
-        dispatch(setGroupSelectedBlocksIds([]))
-      })
+      dispatch(setBlockEditable(null))
     }
   }
 
   return (
-    <>
-      <GridLayoutWrapper
-        ref={builderRef}
-        higlightOnDrop={higlightOnDrop}
-        handleDropNewItem={handleDropNewItem}
-      >
-        <MultipleSelection />
-        <MemoizeLayoutsRender
-          layouts={layoutsKeys}
-          handleHiglightSection={handleHiglightSectionMiddleWare}
-          removeHighlightedElem={removeHighlightedElem}
-          gridRowHeight={gridRowHeight}
-          gridColumnWidth={gridColumnWidth}
-          builderRef={builderRef}
-        />
-      </GridLayoutWrapper>
-    </>
+    <GridLayoutWrapper
+      ref={builderRef}
+      higlightOnDrop={higlightOnDrop}
+      handleDropNewItem={handleDropNewItem}
+    >
+      <MemoizeLayoutsRender
+        layouts={layoutsKeys}
+        handleHiglightSection={handleHiglightSectionMiddleWare}
+        removeHighlightedElem={removeHighlightedElem}
+        gridRowHeight={gridRowHeight}
+        gridColumnWidth={gridColumnWidth}
+        builderRef={builderRef}
+      />
+    </GridLayoutWrapper>
   )
 }
