+++ features/builderSlice.js
@@ -54,12 +54,6 @@ export const builderSlice = createSlice({
     setLoadingData: (state, action) => {
       state.loadingData = action.payload
     },
-    setGroupSelectedBlocksIds: (state, action) => {
-      state.groupSelectedBlocks = action.payload
-    },
-    setIsGroupSelectable: (state, action) => {
-      state.isGroupSelectable = action.payload
-    },
     //This functions are to let the user overwrite DB data
     setTempDBData: (state, action) => {
       //We store here the DB data while the user decides
@@ -160,8 +154,6 @@ export const builderSlice = createSlice({
 export const {
   setBuilderBlocksData,
   setInitialBuilderData,
-  setGroupSelectedBlocksIds,
-  setIsGroupSelectable,
   setTempDBData,
   setUserData,
   setWebsiteId,
@@ -674,17 +666,10 @@ export function findAllChildren(hierarchy, elementDragginId) {
 export const handleDrag =
   (blockPos, newBlockLayout, blockId, gridColumnWidth, gridRowHeight) =>
   (dispatch, getState) => {
-    const state = getState()
-    const groupedBlocks = getGroupSelectedBlocksIds(state)
-    const builderDevice = getBuilderDevice(state)
-    const hierarchy = getHierarchy(state)
-    const children = [
-      ...new Set([
-        ...(findAllChildren(hierarchy, blockId) || []),
-        ...groupedBlocks,
-      ]),
-    ]
-    const layouts = { ...getLayout(state) }
+    const builderDevice = getBuilderDevice(getState())
+    const hierarchy = getHierarchy(getState())
+    const children = [...new Set(findAllChildren(hierarchy, blockId))]
+    const layouts = { ...getLayout(getState()) }
     const updatedHierarchy = getUpdatedHierarchy(
       layouts,
       newBlockLayout,
@@ -733,9 +718,6 @@ export const getHierarchy = (state) => {
   // }
   return getDesktopHierarchy(state)
 }
-export const getGroupSelectedBlocksIds = (state) =>
-  state.builder.groupSelectedBlocks
-export const getIsGroupSelectable = (state) => state.builder.isGroupSelectable
 const getMobileHierarchy = (state) => state.builder.builderData.mobileHierarchy
 const getDesktopHierarchy = (state) => state.builder.builderData.hierarchy
 export const getBlockData = (id) => (state) =>
