+++ landing/hero/index.js
@@ -38,6 +38,7 @@ const Hero = () => {
             135deg
             ,  #43E28E  0%,#506bf0 100%)"
             backgroundClip="text"
+            textFillColor="transparent"
             mx="1rem"
             fontSize={['4rem', '5rem', '5.5rem', '5rem', '6rem']}
           >
@@ -52,6 +53,7 @@ const Hero = () => {
             135deg
             ,  #43E28E  0%,#506bf0 100%)"
             backgroundClip="text"
+            textFillColor="transparent"
             fontSize={['4rem', '5rem', '5.5rem', '5rem', '6rem']}
           >
             Faster
