+++ package.json
@@ -44,7 +44,6 @@
     "react-popper": "^2.2.4",
     "react-quill": "^2.0.0-beta.2",
     "react-redux": "^7.2.2",
-    "react-selecto": "^1.11.1",
     "react-table": "^7.7.0",
     "typed.js": "^2.0.12",
     "unsplash-js": "^7.0.10",
