+++ pages/[resumeId].js
@@ -39,6 +39,7 @@ function Resume({ websiteData, isPublish, resumeId }) {
   if (!isPublish) return <NewPage />
 
   const hasFlurly = hasFlurlyLink(websiteData.blocks)
+  console.log({ hasFlurly })
   return (
     <>
       {hasFlurly && (
