+++ pages/preview/index.js
@@ -31,6 +31,7 @@ function ResumePreview() {
     )
   }
   const hasFlurly = hasFlurlyLink(blocksData.blocks)
+  console.log({ hasFlurly, blocksData })
   return (
     <>
       {hasFlurly && (
